Add files via upload
Changed the amount of control the user has over optional package installs by putting it in a list format and installing them each individually. Added the option to reduce swappiness in the initial sysctl tweaks. Those were for security, but this is more for performance. 
Arch and Antergos don't do this by default and changing the parameter in the kernel is not as easy as doing it in ubuntu, there is no file, so I had to create one.