Add files via upload
made some modifications to the layout of the loops in the script, also added a time date setup routine in the manjaro and antergos setup scripts. This tends to be an issue I run into in arch based systems, it's not always set up for you like it is in Ubuntu.