Add files via upload
Added a new sort command which sorts each individual line of the hosts file in alphabetical order. Runs before uniq -u which will finish the job and finally there is a sed line which removes the useless hash lines on the prefix only leaving the content of the hosts file. This is the best way to get rid of duplicates and I have tested this and it removes almost a good 40,000 lines of unnecessary hosts. I've uploaded the individual hosts updating files on the front page.