Add files via upload
made a few additions to the script, added support for downloading a Paranoid option, a really large option, etc. I also added another prompt towards the end which asks the user if he/she wants to exclude any more sites from their hosts file. May still need some work, but I'm on it as soon as possible, I'm also working on an update for part of the Ubuntu setup script which will automate the process of finding and downloading the correct codecs package. Ubuntu uses window manager or otherwise desktop environment specific packages, but I think I'm on to something. A way that the script no longer needs user input for that. More coming soon.