Add files via upload
Mainly just added the option to limit how much of the package cache the user wants to remove in Manjaro and Antergos. As far as I know this should work the same in both. Consider it a safety protocol. Also the firewalls option now allows the user to deny telnet and ssh if they wish to do so. Soon may allow for custom rules to be added.