Add files via upload
added functionality to check for the distribution and determine which service restart command to use to restart network manager. As of right now, Ubuntu still uses sysvinit for this and arch uses systemd. I have not tested Fedora yet. There are other ways to do this like determining the init process which I will try to implement later. Still testing that.