Add files via upload
Added the first iteration of service management for Ubuntu. Hopefully this is all right and works, I've tested it on the only version of Ubuntu I have and apparently it uses systemd, donno how upstart will work yet. Will test that eventually. Have to go on sourceforge and scour it for an old copy.