Add files via upload
Added an option to install software via the old apt-get based on user choice, much like the one I added to Manjaro and Antergos's script. This has an ever growing list of browsers to install. It also features three main bittorrent clients to choose from right now, but more are on the way.