Add files via upload
Added a restore functionality to the tuneup scripts. This will only restore the home folder assuming it is on the same drive as the system. If you wish to restore the whole system, you'll have to do that in a live environment, I'm going to write a blog article and possibly a script for that in the future.