Add files via upload
Added Driver Info in the part that collects system information, this is a neat thing, it gives users a deeper insight into what is loaded in the kernel on their systems. This is the last update for a while. Possibly the last feature these scripts will get before I update them to the new batch I'm working on.