Add files via upload
Now you get more system information with the system information collection, the Manjaro scripts can work with Antergos, You also get more applications in the list of installable apps. Fixed a few issues with Backup not working. It worked for single machines at least ones I tested it on, but it didn't work for multiple systems, it now asks users which drive they want to use and only backs up the home folder for less confusion. A restore function is coming, but that might be in the next week or so. Fixed some other issues. Will fix more. I know I said I wasn't going to, but this is being ported from the newer scripts so I figured what the heck?