Update Jameshostslist
Removed a dead or no longer active link.
Removed a celebrity news site that kept popping up, annoying me. Perhaps this was rash, someone might use the site. 
Removed Revolt.tv. Aside from the nonsensical popups, it seems to be a legit streaming service.
Removed Flattr as it is a seemingly legit, useful donation platform.
Added businessinsider.com as it is a pain in the @$$ for people using adblockers.